[#1] 파일 2개로 분리
